refactor: rename NativeRudderSdkReactNative to NativeBridge and update imports
