+++ libs/sdk/android/src/main/java/com/rudderstack/react/android/RNRudderSdkModule.java
@@ -19,7 +19,7 @@ import com.rudderstack.android.sdk.core.RudderIntegration;
 import com.rudderstack.android.sdk.core.RudderLogger;
 import com.rudderstack.android.sdk.core.RudderProperty;
 import com.rudderstack.android.sdk.core.RudderMessageBuilder;
-import com.rudderstack.ruddersdkreactnative.NativeRudderSdkReactNativeSpec;
+import com.rudderstack.ruddersdkreactnative.NativeBridgeSpec;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -28,7 +28,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 @ReactModule(name = RNRudderSdkModule.NAME)
-public class RNRudderSdkModule extends NativeRudderSdkReactNativeSpec {
+public class RNRudderSdkModule extends NativeBridgeSpec {
 
     public static final String NAME = "RNRudderSdkModule";
     private final ReactApplicationContext reactContext;
